# Conflicts:
#	.gitignore
#	LOCAL_SCOPE.md
#	assets_local/dungeons/dungeons.txt
#	docs/building-apps.md
#	docs/design-language.md
#	docs/dungeons.md
#	host-tests/battleship/run.sh
#	host-tests/connections/run.sh
#	host-tests/connections/test_connections.cpp
#	host-tests/dungeon/run.sh
#	host-tests/hackernews/fixtures.h
#	host-tests/hackernews/test_hackernews.cpp
#	host-tests/insider/run.sh
#	host-tests/jaipur/run.sh
#	host-tests/solitaire/run.sh
#	host-tests/ui/run.sh
#	host-tests/ui/test_ui.cpp
#	platformio.sim.ini
#	scripts_local/check.sh
#	scripts_local/lib-sim.sh
#	scripts_local/sim_catchup.py
#	scripts_local/sync.sh
#	src/apps_local/Shelf.cpp
#	src/apps_local/ShelfFolderActivity.cpp
#	src/apps_local/ShelfFolderActivity.h
#	src/apps_local/battleship/BattleshipActivity.cpp
#	src/apps_local/battleship/BattleshipScreens.cpp
#	src/apps_local/chess/ChessActivity.cpp
#	src/apps_local/chess/ChessEngine.cpp
#	src/apps_local/chess/ChessScreens.cpp
#	src/apps_local/connections/ConnectionsActivity.cpp
#	src/apps_local/dungeon/DungeonActivity.cpp
#	src/apps_local/dungeon/DungeonActivity.h
#	src/apps_local/dungeon/DungeonCore.h
#	src/apps_local/dungeon/DungeonPuzzles.h
#	src/apps_local/dungeon/DungeonScreens.cpp
#	src/apps_local/hackernews/HackerNewsActivity.cpp
#	src/apps_local/hackernews/HackerNewsActivity.h
#	src/apps_local/hackernews/HackerNewsCore.cpp
#	src/apps_local/hackernews/HackerNewsCore.h
#	src/apps_local/hackernews/HackerNewsScreens.cpp
#	src/apps_local/hackernews/HackerNewsScreens.h
#	src/apps_local/insider/InsiderActivity.cpp
#	src/apps_local/insider/InsiderScreens.cpp
#	src/apps_local/jaipur/JaipurActivity.cpp
#	src/apps_local/link/LinkActivity.cpp
#	src/apps_local/link/LinkRadio.cpp
#	src/apps_local/player/PlayerActivity.cpp
#	src/apps_local/solitaire/SolitaireActivity.cpp
#	src/apps_local/study/StudyActivity.cpp
#	src/apps_local/study/StudyFonts.cpp
#	src/apps_local/study/StudyFonts.h
#	src/apps_local/ui/ToyboxIcons.h
#	src/apps_local/ui/ToyboxScreen.h
#	src/apps_local/ui/ToyboxTokens.h
#	src/apps_local/ui/fonts/reading_serif_11.h
#	src/apps_local/ui/fonts/reading_serif_14.h
#	src/apps_local/ui/fonts/reading_serif_bold_12.h
#	src/apps_local/ui/fonts/reading_serif_bold_16.h
#	src/apps_local/ui/fonts/reading_serif_bold_18.h
#	src/components/icons/shelfIcons.h
#	tools_local/gen_dungeons.py
#	tools_local/gen_toybox_icons.sh
#	tools_local/icons.txt

+++ src/apps_local/chess/ChessEngine.cpp
@@ -102,7 +102,10 @@ const int8_t* tableFor(const uint8_t type) {
 int moveScore(const Position& position, const Move& move) {
   int score = 0;
   if ((move.flags & FlagCapture) != 0) {
-    const uint8_t victim = (move.flags & FlagEnPassant) != 0 ? Pawn : pieceType(position.squares[move.to]);
+    // static_cast keeps both arms uint8_t: Piece is an enum and pieceType()
+    // returns the plain type, which -Wextra flags as a mixed conditional.
+    const uint8_t victim =
+        (move.flags & FlagEnPassant) != 0 ? static_cast<uint8_t>(Pawn) : pieceType(position.squares[move.to]);
     const uint8_t attacker = pieceType(position.squares[move.from]);
     score += 10000 + kPieceValue[victim] * 10 - kPieceValue[attacker];
   }
